+++ src/Settings/SettingsInfo.hpp
@@ -20,6 +20,7 @@
 
 #include <QCoreApplication>
 #include <QVariant>
+#include <functional>
 
 class SettingsInfo
 {
@@ -30,7 +31,7 @@ class SettingsInfo
     {
         QString name, desc, type, ui, tip, help;
         bool requireAllDepends; // false for one of the depends, true for all depends
-        QStringList depends;
+        QList<QPair<QString, std::function<bool(const QVariant &)>>> depends;
         QVariant def;
         QVariant param;
         QList<SettingInfo> child;