+++ doc/CHANGELOG.md
@@ -16,6 +16,7 @@ With the portable version, you can easily store it in something like a USB disk,
 - Auto-save time interval.
 - Now you can auto-save the current session, besides saving it when the application exists. (#437 and #442)
 - Now the width of the input/output/expected of each test case is adjustable, and the maximum height of a test case can be set in the preferences. (#414 and #444)
+- Now you can use proxy to check for updates. (#448)
 
 ### Fixed